Add unit tests for PathStorage

+++ src/main/java/ua/foxminded/moldavets/project/storage/PathStorage.java
@@ -1,5 +1,6 @@
 package ua.foxminded.moldavets.project.storage;
 
+import ua.foxminded.moldavets.project.exception.ExistStorageException;
 import ua.foxminded.moldavets.project.exception.NotExistStorageException;
 import ua.foxminded.moldavets.project.exception.StorageException;
 import ua.foxminded.moldavets.project.model.Resume;
@@ -40,44 +41,72 @@ public class PathStorage implements FileStorage<Path> {
 
     @Override
     public void save(Resume resume, Path searchKey) {
-        try {
-            serializer.subWrite(resume, new BufferedOutputStream(new FileOutputStream(getSearchKey(resume.getUuid()).toFile())));
-            System.out.println(resume.getUuid() + " saved successfully");
-        } catch (IOException e) {
-            throw new StorageException("Couldn't save resume ", resume.getUuid(), e);
+
+        if (resume != null && searchKey != null) {
+            if (!searchKey.toFile().getName().trim().isEmpty() && !resume.getUuid().trim().isEmpty()) {
+                if (!isExist(searchKey)) {
+                    try {
+                        serializer.subWrite(resume, new BufferedOutputStream(new FileOutputStream(getSearchKey(resume.getUuid()).toFile())));
+                        System.out.println(resume.getUuid() + " saved successfully");
+                    } catch (IOException e) {
+                        throw new StorageException("Couldn't save resume ", resume.getUuid(), e);
+                    }
+                } else {
+                    throw new ExistStorageException(resume.getUuid());
+                }
+            } else {
+                throw new StorageException("Search key or resume cannot be empty", resume.getUuid());
+            }
+        } else {
+            throw new NullPointerException("Resume or search key cannot be null");
         }
     }
 
     @Override
     public void update(Resume resume, Path searchKey) {
-        if (isExist(searchKey)) {
-            try {
-                serializer.subWrite(resume, new BufferedOutputStream(new FileOutputStream(getSearchKey(resume.getUuid()).toFile())));
-                System.out.println(resume.getUuid() + " updated successfully");
-            } catch (IOException e) {
-                throw new StorageException("Couldn't update resume ", resume.getUuid(), e);
+
+        if (resume != null && searchKey != null) {
+            if (isExist(searchKey)) {
+                try {
+                    serializer.subWrite(resume, new BufferedOutputStream(new FileOutputStream(getSearchKey(resume.getUuid()).toFile())));
+                    System.out.println(resume.getUuid() + " updated successfully");
+                } catch (IOException e) {
+                    throw new StorageException("Couldn't update resume ", resume.getUuid(), e);
+                }
+            } else {
+                throw new NotExistStorageException(resume.getUuid());
             }
         } else {
-            throw new NotExistStorageException(resume.getUuid());
+            throw new NullPointerException("Resume or search key cannot be null");
         }
     }
 
     @Override
     public void delete(Path searchKey) {
-        try {
-            Files.delete(searchKey);
-            System.out.printf("File %s deleted successful", searchKey.toFile().getName());
-        } catch (IOException e) {
-            throw new StorageException("Path delete exception", null, e);
+
+        if (searchKey != null) {
+            try {
+                Files.delete(searchKey);
+                System.out.printf("File %s deleted successful", searchKey.toFile().getName());
+            } catch (IOException e) {
+                throw new StorageException("Path delete exception", null, e);
+            }
+        } else {
+            throw new NullPointerException("Path cannot be null");
         }
     }
 
     @Override
     public void clear() {
-        try {
-            Files.list(directory).forEach(this::delete);
-        } catch (IOException e) {
-            throw new StorageException("Path delete exception", null, e);
+
+        if (getSize() != 0 ) {
+            try {
+                Files.list(directory).forEach(this::delete);
+            } catch (IOException e) {
+                throw new StorageException("Path delete exception", null, e);
+            }
+        } else {
+            throw new StorageException("Storage already empty", null);
         }
     }
 
@@ -86,13 +115,14 @@ public class PathStorage implements FileStorage<Path> {
         String[] files = directory.toFile().list();
 
         if(files == null) {
-            throw new StorageException("Storage null exception", null);
+            throw new StorageException("Storage is empty", null);
         }
         return files.length;
     }
 
     @Override
     public boolean isExist(Path searchKey) {
+        System.out.println(searchKey);
         return Files.exists(searchKey);
     }
 }
